Move *.glsl file loading into a global require script
